requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
